// src/app/model/personality-types.enum.ts
export enum PersonalityTypes {
    THINKING_EXTROVERT = 'THINKING EXTROVERT',
    THINKING_INTROVERT = 'THINKING INTROVERT',
    SENTIMENTAL_EXTROVERT = 'SENTIMENTAL EXTROVERT',
    SENTIMENTAL_INTROVERT = 'SENTIMENTAL INTROVERT',
    SENSITIVE_EXTROVERT = 'SENSITIVE EXTROVERT',
    SENSITIVE_INTROVERT = 'SENSITIVE INTROVERT',
    INTUITIVE_EXTROVERT = 'INTUITIVE EXTROVERT',
    INTUITIVE_INTROVERT = 'INTUITIVE INTROVERT',
}